feature: added a HtmlHint config file
Added a .htmlhintrc because the default wants spaces for indentation. Since htmlhint doesn't merge the custom and the default config file, I copied one from the internet (I couldn't find the normal default). Feel free to make changes if you want to.

(The file doesn't have the .json extension because htmlhint didn't find it when it did)